Move pipeline inside
