+++ nemo_skills/pipeline/run_training.py
@@ -21,8 +21,7 @@ from pathlib import Path
 # adding nemo_skills to python path to avoid requiring installation
 sys.path.append(str(Path(__file__).absolute().parents[1]))
 
-from launcher import CLUSTER_CONFIG, NEMO_SKILLS_CODE, fill_env_vars, launch_job
-
+from nemo_skills.pipeline.utils import CLUSTER_CONFIG, NEMO_SKILLS_CODE, fill_env_vars, launch_job
 from nemo_skills.utils import setup_logging
 
 # note that we are using custom config nemo_skills/finetuning/sft_config.py
